Refactor login/signup system
login and signup now use methods from the new UserController class.

For convenience, the login button on the signup modal will now log you directly into the app rather than taking you to the screen to fill out your information again.

Changed the login page login button to look the same as the rest of the buttons on the site.